Revert "Create CODEOWNERS"
This reverts commit c3e935da308dff043fddb8d47841e3d2394ac91a.

I inconsiderately decided to add myself to be notified for reviews.

However there are far more cases of dependency upgrades than actual
contributors, and this would just lead to a ton of noise in my
feed. On top of that, I added this in the wrong place. The CODEOWNER
addition will be reverted—while I am and for the foreseeable future
will be maintaining this repo, it isn't worth it to be notified about
every update.